python/requirements.txt
# onnxruntime
onnxruntime-gpu
opencv-python